requirements.txt
Django
gunicorn
whitenoise
python-decouple